Module Drm.Buffer

Memory holding pixel data.

type id = [ `Buffer ] Id.t

A GEM buffer handle.

From Linux's drm.h:

GEM handles are not reference-counted by the kernel. User-space is responsible for managing their lifetime. For example, if user-space imports the same memory object twice on the same DRM file description, the same GEM handle is returned by both imports, and user-space needs to ensure DRM_IOCTL_GEM_CLOSE is performed once only. The same situation can happen when a memory object is allocated, then exported and imported again on the same DRM file description. The DRM_IOCTL_MODE_GETFB2 IOCTL is an exception and always returns fresh new GEM handles even if an existing GEM handle already refers to the same memory object before the IOCTL is performed.

val close : Device.t -> id -> unit

close id calls DRM_IOCTL_GEM_CLOSE (see warning comment on id).

module Dumb : sig ... end

A slow and inefficient buffer in host memory for testing, etc.