Module Drm.Kms

Kernel mode-setting.

This API is used to configure physical monitors (screen resolution, framebuffers, etc).

See the Linux Kernel Mode Setting documentation for more information.

type uint32 := Unsigned.UInt32.t

module type BITSET = sig ... end

module Rect : sig ... end

module Blob : sig ... end

An untyped string of bytes.

module Mode_info : sig ... end

Screen mode resolution and timing information.

module Sub_pixel : sig ... end

How pixels are physically arranged on a particular monitor.

module Property : sig ... end

A property can be used to read or write extra settings on an object (resource).

module Properties : sig ... end

Drivers can add extra properties to objects, discoverable at runtime.

module Connector : sig ... end

A physical connector used to attach a monitor.

module Fb : sig ... end

A framebuffer manages the inputs to a Crtc.

module Crtc : sig ... end

A CRT Controller.

module Plane : sig ... end

A plane wraps a Buffer with some extra metadata.

module Encoder : sig ... end

Encoders reads pixel data from a Crtc and output it in a suitable format for a Connector.

module Resources : sig ... end

The set of resources provided by the hardware (excluding planes; see Plane.list).

module Atomic_req : sig ... end

Atomic requests.

module Lease : sig ... end

Sharing access with other processes.