Module Kms.Connector

A physical connector used to attach a monitor.

module Connection : sig ... end

User-space should first try to enable Connected connectors and ignore other connectors. If there are no Connected connectors, user-space should then try to probe and enable Unknown_connection connectors.

module Type : sig ... end

type id = [ `Connector ] Id.t

type t = {

connector_id : id;

encoder_id : [ `Encoder ] Id.t option;

connector_type : Type.t;

connector_type_id : int;

connection : Connection.t;

mm_width : int;

mm_height : int;

subpixel : Sub_pixel.t;

modes : Mode_info.t list;

props : Properties.Values.raw;

encoders : [ `Encoder ] Id.t list;

}

val get : Device.t -> id -> t

Retrieve all information about the connector. This will do a forced probe on the connector to retrieve remote information such as EDIDs from the display device.

val get_current : Device.t -> id -> t

Retrieve current information, i.e the currently active mode and encoder, about the connector. This will not do any probing on the connector or remote device, and only reports what is currently known. For the complete set of modes and encoders associated with the connector use get which will do a probe to determine any display link changes first.

val id : t -> id

val pp_modes : Mode_info.t list Fmt.t

val pp_name : t Fmt.t

e.g. "DP-1"

val pp : t Fmt.t

Properties

type 'a property = ([ `Connector ], 'a) Property.t

val get_properties : Device.t -> id -> [ `Connector ] Properties.Values.t

val crtc_id : [ `Crtc ] Id.t option property

val edid : Blob.id option property

val tile : Blob.id option property

val link_status : [ `Good | `Bad | `Unknown of Property.raw_value ] property

val non_desktop : bool property