Module Connector.Connection

User-space should first try to enable Connected connectors and ignore other connectors. If there are no Connected connectors, user-space should then try to probe and enable Unknown_connection connectors.

type t =

| Connected
(*
The connector has a sink plugged in
*)

| Disconnected

| Unknown_connection

val pp : t Fmt.t