Module Kms.Encoder

Encoders reads pixel data from a Crtc and output it in a suitable format for a Connector.

The Linux documentation says that exposing encoders to user-space was a design mistake. With the newer atomic API, you can use Connector.crtc_id to get the CRTC directly.

module Type : sig ... end

type id = [ `Encoder ] Id.t

type t = {

encoder_id : id;

encoder_type : Type.t;

crtc_id : Crtc.id option;

possible_crtcs : int;

possible_clones : int;

}

val get : Device.t -> id -> t

val id : t -> id

val pp : t Fmt.t