Module Mode_info.Flags

include BITSET

type t = private uint32

val empty : t

empty is the empty set (i.e. 0).

val (+) : t -> t -> t

a + b is the union of a and b (i.e. logor).

val mem : t -> t -> bool

mem flag x checks if flag is set in x. (i.e. if x land flag = flag)

val of_uint32 : uint32 -> t

val pp : t Fmt.t

val phsync : t

val nhsync : t

val pvsync : t

val nvsync : t

val interlace : t

val dblscan : t

val csync : t

val pcsync : t

val ncsync : t

val hskew : t

val bcast : t

val pixmux : t

val dblclk : t

val clkdiv2 : t