Module Mode_info.Type

include BITSET

type t = private uint32

val empty : t

empty is the empty set (i.e. 0).

val (+) : t -> t -> t

a + b is the union of a and b (i.e. logor).

val mem : t -> t -> bool

mem flag x checks if flag is set in x. (i.e. if x land flag = flag)

val of_uint32 : uint32 -> t

val pp : t Fmt.t

val builtin : t

val clock_c : t

val crtc_c : t

val preferred : t

val default : t

val userdef : t

val driver : t