Module Kms.Properties

Drivers can add extra properties to objects, discoverable at runtime.

module Type : sig ... end

type 'a metadata

Metadata about extra properties available on an object.

val object_id : 'a metadata -> 'a Id.t

val lookup_property : 'a metadata -> ('a, _) Property.t -> Property.Info.t option

lookup_name metadata p returns information about the property p.

val set_value : Device.t -> 'a metadata -> ('a, 'v) Property.t -> 'v -> unit

set_value dev t p v sets the value of p to v on the object t.

module Values : sig ... end