Module type Kms.BITSET

type t = private uint32

val empty : t

empty is the empty set (i.e. 0).

val (+) : t -> t -> t

a + b is the union of a and b (i.e. logor).

val mem : t -> t -> bool

mem flag x checks if flag is set in x. (i.e. if x land flag = flag)

val of_uint32 : uint32 -> t

val pp : t Fmt.t